% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat.R
\name{combat}
\alias{combat}
\title{Combatting batch effects when combining batches of gene expression microarray 
data}
\usage{
combat(
  dat,
  batch,
  mod = NULL,
  nuisance.mod = NULL,
  train = NULL,
  resid = FALSE,
  eb = TRUE,
  parametric = TRUE,
  mean.only = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dat}{A \emph{p x n} matrix (or object coercible by \link[base]{as.matrix}
to a numeric matrix) of observations where \emph{p} is the number of
features and \emph{n} is the number of subjects.}

\item{batch}{Factor (or object coercible by \link[base]{as.factor} to a 
factor) designating batch IDs.}

\item{mod}{An optional design matrix to preserve, usually output of 
\link[stats]{model.matrix}.}

\item{nuisance.mod}{An optional design matrix to regress out, usually output 
of \link[stats]{model.matrix} without intercept.}

\item{train}{Optional logical vector specifying subset of observations used
to estimate coefficients.}

\item{resid}{Whether to leave intercept and covariates regressed out.}

\item{eb}{If \code{TRUE}, uses ComBat model with empirical Bayes.}

\item{parametric}{If \code{TRUE}, uses parametric updates.}

\item{mean.only}{If \code{TRUE}, ComBat step does not harmonize variance.}

\item{verbose}{Whether additional details are printed to console.}
}
\value{

}
\description{
\code{combat} is a modified version of the ComBat code written by 
Jean-Philippe Fortin available at 
\url{https://github.com/Jfortin1/ComBatHarmonization/}. The function
harmonizes the mean and variance of observations across sites under an
empirical Bayes framework. \code{combat} additionally includes options
to output residualized observations, estimate coefficients using a training
subset, and regress out unwanted confounders.
}
\examples{

}
\seealso{
Modification to ComBat to regress out unwanted confounders 
proposed by Wachinger et al. (2020), \url{https://arxiv.org/abs/2002.05049}
}
