% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat2.R
\name{combat2}
\alias{combat2}
\title{Combatting batch effects when combining batches of gene expression microarray 
data}
\usage{
combat2(
  dat,
  batch,
  mod = NULL,
  var.mod = NULL,
  nuisance.mod = NULL,
  train = NULL,
  resid = FALSE,
  eb = TRUE,
  parametric = TRUE,
  mean.only = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{}

\item{batch}{}

\item{mod}{An optional design matrix to preserve, usually output of 
\link[stats]{model.matrix}.}

\item{var.mod}{An optional design matrix to preserve in the variance, usually
output of \link[stats]{model.matrix}.}

\item{nuisance.mod}{An optional design matrix to regress out, usually output 
of \link[stats]{model.matrix} without intercept.}

\item{train}{}

\item{resid}{}

\item{eb}{}

\item{parametric}{}

\item{mean.only}{}

\item{verbose}{}
}
\value{

}
\description{
\code{combat} is a modified version of the ComBat code written by 
Jean-Philippe Fortin available at 
\url{https://github.com/Jfortin1/ComBatHarmonization/}. The function
harmonizes the mean and variance of observations across sites under an
empirical Bayes framework. \code{combat} additionally includes options
to output residualized observations, estimate coefficients using a training
subset, and regress out unwanted confounders. \code{combat2} further includes
options to control for covariate effects in variance using 
\link[lmvar]{lmvar}
}
\examples{

}
\seealso{
Modification to ComBat to regress out unwanted confounders 
proposed by Wachinger et al. (2020), \url{https://arxiv.org/abs/2002.05049}
}
