% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_modded.R
\name{combat_modded}
\alias{combat_modded}
\title{Combatting batch effects when combining batches of gene expression microarray data}
\usage{
combat_modded(
  dat,
  batch,
  mod = NULL,
  eb = TRUE,
  verbose = TRUE,
  parametric = TRUE,
  mean.only = FALSE,
  resid = FALSE,
  train = NULL
)
}
\arguments{
\item{dat}{}

\item{batch}{}

\item{mod}{}

\item{eb}{}

\item{verbose}{}

\item{parametric}{}

\item{mean.only}{}

\item{resid}{}

\item{train}{}
}
\value{

}
\description{
Combatting batch effects when combining batches of gene expression microarray data
}
