% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covbat.R
\name{covbat}
\alias{covbat}
\title{Correcting Covariance Batch Effects: CovBat}
\usage{
covbat(
  x,
  bat,
  mod = NULL,
  percent.var = 0.95,
  n.pc = NULL,
  train = NULL,
  mean.only = FALSE,
  std.var = TRUE,
  resid = FALSE,
  eb = TRUE,
  parametric = TRUE,
  score.eb = FALSE,
  score.parametric = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A \emph{p x n} matrix (or object coercible by \link[base]{as.matrix}
to a numeric matrix) of observations where \emph{p} is the number of
features and \emph{n} is the number of subjects.}

\item{bat}{A factor (or object coercible by \link[base]{as.factor} to a 
factor) designating batch IDs.}

\item{mod}{An optional design matrix to preserve, usually output of 
\link[stats]{model.matrix}.}

\item{percent.var}{A numeric. The number of harmonized principal component 
scores is selected to explain this proportion of the variance.}

\item{n.pc}{An optional numeric. If specified, this number of principal
component scores is harmonized. Overrides \code{percent.var}.}

\item{train}{An optional vector specifying subset of observations used to 
estimate coefficients.}

\item{mean.only}{If \code{TRUE}, ComBat step does not harmonize variance.}

\item{std.var}{If \code{TRUE}, scales variances to be equal to 1 before PCA.}

\item{resid}{Whether to leave intercept and covariates regressed out.}

\item{eb}{If \code{TRUE}, uses ComBat model with empirical Bayes for mean
and variance harmonization.}

\item{parametric}{If \code{TRUE}, uses parametric adjustments}

\item{score.eb}{If \code{TRUE}, uses ComBat model with empirical Bayes for
harmonization of scores.}

\item{score.parametric}{If \code{TRUE}, uses  parametric ComBat is used
for harmonization of scores.}

\item{verbose}{Whether additional details are printed to console.}
}
\value{
\code{covbat} returns a list containing the following components:
\item{dat.covbat}{harmonized data as a matrix with same dimensions as 
   \code{x}}
}
\description{
\code{covbat} harmonizes covariance of observations across sites. It
first applies ComBat to harmonize mean and variance then adjusts variance of
PC scores to harmonize covariance.
}
